+++ packages/datafeed/src/equicast_datafeed/__init__.py
@@ -2,7 +2,8 @@
 
 from equicast_datafeed.client import DatafeedClient
 from equicast_datafeed.exceptions import DatafeedError
+from equicast_datafeed.rounding import DECIMAL_PRECISION, round_value
 
 __version__ = "0.1.0"
 
-__all__ = ["DatafeedClient", "DatafeedError"]
+__all__ = ["DatafeedClient", "DatafeedError", "DECIMAL_PRECISION", "round_value"]